requirements.txt
Flask>=3.0.0
numpy>=1.26.0
opencv-python>=4.10.0
python-dotenv>=1.0.0
groq>=0.11.0
